requirements.txt
pandas>=1.3.0
numpy>=1.16.5
backpyf==0.9.2.038
binance-futures-connector==4.1.0